requirements-travis.txt
gunicorn==20.0.4
Django==3.1.1
django-heroku==0.3.1
django-allauth==0.42.0
python-dotenv==0.14.0
requests==2.24.0
django-crispy-forms
